pyproject.toml
[tool.poetry]
name = "sjandroidbilling"
version = "2022.1.0"
description = "port google play billing (android) to python"
authors = ["kengoon <kengoon19@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
pyjnius = "^1.4.2"
kvdroid = {url = "https://github.com/kvdroid/Kvdroid/archive/refs/heads/master.zip"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
